fixed many errors flagged by the 10.6 xcode compiler analysis tools
git-svn-id: svn://orca.physics.unc.edu/Orca@2417 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Misc/SynthesizeSingleton.h
@@ -36,7 +36,7 @@ return self; \
 return self; \
 } \
 \
-- (unsigned)retainCount \
+- (NSUInteger)retainCount \
 { \
 return 0xffffffff; \
 } \
@@ -87,7 +87,7 @@ return self; \
 return self; \
 } \
 \
-- (unsigned)retainCount \
+- (NSUInteger)retainCount \
 { \
 return 0xffffffff; \
 } \
